docs: note analytics stub and future replacement
