--- packages/analytics/.prettierignore
@@ -1,1 +1,0 @@
-src/summary.ts
