+++ packages/analytics/README.md
@@ -1,0 +1,10 @@
+# @prism-apex-tool/analytics (stub)
+
+Temporary no-op analytics facade used to keep CI green.
+
+- `trackEvent(name, props)` – no-op
+- `trackError(error, context)` – no-op
+- `meter(name, value, tags)` – no-op
+- `createAnalyticsScope(scope)` – returns the same no-op fns
+
+Replace with a real implementation in a later telemetry PR.
