+++ packages/analytics/package.json
@@ -3,24 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "type": "module",
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "lint": "eslint . --ext .ts",
-    "test": "vitest --run --passWithNoTests",
-    "typecheck": "tsc -p tsconfig.json --noEmit",
-    "format": "prettier -w .",
-    "format:check": "prettier -c ."
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs",
+      "types": "./dist/index.d.ts"
+    }
   },
-  "dependencies": {},
+  "scripts": {
+    "build": "tsup src/index.ts --format esm,cjs --dts",
+    "test": "vitest run --passWithNoTests",
+    "lint": "eslint ."
+  },
   "devDependencies": {
-    "eslint": "^9.10.0",
-    "@eslint/js": "^9.10.0",
-    "typescript-eslint": "^8.7.0",
-    "prettier": "^3.3.3"
+    "tsup": "^8.0.1",
+    "typescript": "^5.4.0",
+    "vitest": "^1.6.0",
+    "@types/node": "^20.11.0"
   }
 }
