+++ pnpm-workspace.yaml
@@ -3,3 +3,4 @@ packages:
   - 'packages/*'
   - 'packages/rules-apex'
   - 'packages/accounts'
+  - 'packages/analytics'
